requirements.txt
Flask==2.3.3
flask-cors==4.0.0
googletrans==4.0.0rc1
langdetect==1.0.9
python-dotenv==1.0.0
twilio==8.10.0          # optional, for real WhatsApp integration